assets/search/requirements.txt
fabric==0.9.3
-e git+https://github.com/django/django.git@1.2.3#egg=django
django-debug-toolbar
python-dateutil
pymongo
